Create TerminalModesAdder transformer
